# Optimize async Promise