Trade-off increased memory to remove a branch, resulting in faster synchronous completions (asynchronous completion times are negligible).

develop

|       Method | Pending |       Mean | Allocated | Survived |
|------------- |-------- |-----------:|----------:|---------:|
| AsyncAwait |   False |   334.0 ns |         - |        - |
| AsyncAwait |    True | 1,602.5 ns |         - |    624 B |

PR

|       Method | Pending |       Mean | Allocated | Survived |
|------------- |-------- |-----------:|----------:|---------:|
| AsyncAwait |   False |   304.0 ns |         - |        - |
| AsyncAwait |    True | 1,611.2 ns |         - |    648 B |